requirements.txt
cachetools
flask-pydantic
functions-framework==3.*
mlflow
pandas
scikit-learn
